# docs/openapi/clipmato-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clipmato Public API",
    "summary": "Versioned public API contracts for uploads, progress, records, scheduling, and publishing.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/progress/{record_id}": {
      "get": {
        "tags": [
          "Public API"
        ],
        "summary": "Get Progress",
        "description": "Return the current progress state for a record.",
        "operationId": "get_progress_api_v1_progress__record_id__get",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/record/{record_id}/schedule": {
      "post": {
        "tags": [
          "Public API"
        ],
        "summary": "Schedule Record",
        "description": "Schedule a record and persist provider publish jobs.",
        "operationId": "schedule_record_api_v1_record__record_id__schedule_post",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJobUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/upload": {
      "post": {
        "tags": [
          "Public API"
        ],
        "summary": "Upload",
        "description": "Accept an upload using the versioned public contract.",
        "operationId": "upload_api_v1_upload_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_v1_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency key conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_api_v1_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "remove_silence": {
            "type": "boolean",
            "title": "Remove Silence",
            "default": false
          },
          "selected_project_presets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Project Presets",
            "default": []
          },
          "project_name": {
            "type": "string",
            "title": "Project Name",
            "default": ""
          },
          "project_summary": {
            "type": "string",
            "title": "Project Summary",
            "default": ""
          },
          "project_topics": {
            "type": "string",
            "title": "Project Topics",
            "default": ""
          },
          "project_prompt_prefix": {
            "type": "string",
            "title": "Project Prompt Prefix",
            "default": ""
          },
          "project_prompt_suffix": {
            "type": "string",
            "title": "Project Prompt Suffix",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_v1_upload_post"
      },
...
      "MachineError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "status",
          "message"
        ],
        "title": "MachineError",
        "description": "Standard machine-readable error payload for public API responses."
      },
      "MachineErrorEnvelope": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "error": {
            "$ref": "#/components/schemas/MachineError"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "error"
        ],
        "title": "MachineErrorEnvelope",
        "description": "Envelope for error responses that also echoes the correlation ID."
      },
      "ProgressStatusResponse": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "progress"
        ],
        "title": "ProgressStatusResponse"
      },
...
      "PublishJobModel": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "scheduled_for": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled For"
          },
          "next_attempt_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Attempt At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "attempt_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Count"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "privacy_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Status"
          },
          "remote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Id"
          },
          "remote_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Url"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "type": "object",
        "title": "PublishJobModel"
      },
      "PublishJobUpdateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "schedule_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Time"
          },
          "publish_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Publish Targets"
          },
          "publish_jobs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PublishJobModel"
            },
            "type": "object",
            "title": "Publish Jobs"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PublishJobUpdateResponse"
      },
...
      "ScheduleRecordRequest": {
        "properties": {
          "schedule_time": {
            "type": "string",
            "title": "Schedule Time"
          },
          "publish_targets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Publish Targets"
          },
          "youtube_privacy_status": {
            "type": "string",
            "title": "Youtube Privacy Status",
            "default": "private"
          }
        },
        "type": "object",
        "required": [
          "schedule_time"
        ],
        "title": "ScheduleRecordRequest"
      },
...
      "UploadAcceptedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "UploadAcceptedResponse"
      },
...
    }
  }
}